Switching from Slop to Thor for the CLI
