+++ bin/test_gems
@@ -4,7 +4,7 @@ require 'deep_cover'
 
 def gem_list
   return %w[erubis thor i18n builder tzinfo diff-lcs rack-test tilt multipart-post diff-lcs
-            rubygems-bundler method_source daemons hike multi_json rails-deprecated_sanitizer slop
+            rubygems-bundler method_source daemons hike multi_json rails-deprecated_sanitizer
             rspec-support thread_safe rspec-core rspec-expectations rspec-mocks rspec mini_portile
             multi_json rspec-support rspec-expectations rspec-mocks mail nokogiri rack rspec unf
             json thread_safe rspec-core json rack-protection sdoc docile faraday multi_xml coderay
