+++ deep_cover.gemspec
@@ -29,7 +29,7 @@ Gem::Specification.new do |spec|
   # CLI
   spec.add_runtime_dependency 'bundler'
   spec.add_runtime_dependency 'highline'
-  spec.add_runtime_dependency 'slop', '~> 4.0'
+  spec.add_runtime_dependency 'thor'
   spec.add_runtime_dependency 'with_progress'
 
   ### Dev dependencies
