+++ lib/deep_cover/cli/exec.rb
@@ -1,26 +1,42 @@
+#!/usr/bin/env ruby
 # frozen_string_literal: true
 
+require_relative '../cli'
+
 module DeepCover
-  module CLI
-    class Exec
-      class Option
-        def keep_file_descriptors?
-        end
+  class CLI
+    # Stop parsing and treat everything as positional argument if
+    # we encounter an unknown option or a positional argument.
+    # `check_unknown_options!` (defined in cli.rb) happens first,
+    # so we just stop on a positional argument.
+    stop_on_unknown_option! :exec
+
+    desc 'exec [OPTIONS] COMMAND TO RUN', 'Execute the command with coverage activated'
+    option '--output', desc: 'output folder', type: :string, default: DeepCover.config.output, aliases: '-o'
+    option '--reporter', desc: 'reporter to use', type: :string, default: DeepCover.config.reporter
+    option '--open', desc: 'open the output coverage', type: :boolean, default: CLI_DEFAULTS[:open]
+
+    def exec(*command_parts)
+      if command_parts.empty?
+        warn set_color('`exec` needs a command to run', :red)
+        exit(1)
       end
 
-      def initialize(argv, **options)
-        @argv = argv
-        @options = options
-      end
+      require 'yaml'
+      env_var = {'DEEP_COVER' => 't',
+                 'DEEP_COVER_OPTIONS' => YAML.dump(processed_options.slice(*DEFAULTS.keys)),
+      }
 
-      def run
-        require 'yaml'
-        require 'deep_cover/backports'
-        env_var = {'DEEP_COVER' => 't',
-                   'DEEP_COVER_OPTIONS' => YAML.dump(@options.slice(*DEFAULTS.keys)),
-                  }
-
-        system(env_var, *@argv)
+      # Clear inspiration from Bundler's kernel_exec
+      # https://github.com/bundler/bundler/blob/d44d803357506895555ff97f73e60d593820a0de/lib/bundler/cli/exec.rb#L50
+      begin
+        Kernel.exec(env_var, *command_parts)
+      rescue Errno::EACCES, Errno::ENOEXEC
+        warn set_color("not executable: #{command_parts.first}", :red)
+        exit 126 # Default exit code for that
+      rescue Errno::ENOENT
+        warn set_color("command not found: #{command_parts.first}", :red)
+        exit 127 # Default exit code for that
       end
     end
   end
