+++ spec/cli_spec.rb
@@ -16,7 +16,7 @@ module DeepCover
 
     describe 'deep-cover exec' do
       let(:options) { '' }
-      let(:command) { "cd spec/code_fixtures/#{path} && ../../../exe/deep-cover -o=false #{options} exec rake" }
+      let(:command) { "cd spec/code_fixtures/#{path} && ../../../exe/deep-cover exec -o=false #{options} rake" }
       subject { output }
       describe 'for a simple gem' do
         let(:path) { 'covered_trivial_gem' }
@@ -27,7 +27,7 @@ module DeepCover
       end
 
       describe 'for a command with options' do
-        let(:command) { %{cd exe; ./deep-cover -o=false exec ruby -I../lib -I../core_gem/lib -e 'require "deep-cover"; puts :hello'} }
+        let(:command) { %{cd exe; ./deep-cover exec -o=false ruby -I../lib -I../core_gem/lib -e 'require "deep-cover"; puts :hello'} }
         it do
           should include 'hello'
           should =~ /No HTML generated/
@@ -35,9 +35,10 @@ module DeepCover
       end
     end
 
-    describe 'The output of deep-cover' do
+    describe 'The output of deep-cover clone' do
       let(:options) { '' }
-      let(:command) { "cd exe; ./deep-cover ../spec/code_fixtures/#{path} -o=false --reporter=istanbul --no-bundle #{options}" }
+      let(:extra_args) { '' }
+      let(:command) { "cd exe; ./deep-cover clone -o=false --reporter=istanbul --no-bundle #{options} ../spec/code_fixtures/#{path}" }
       subject { output }
       describe 'for a simple gem' do
         let(:path) { '../../core_gem/spec/code_fixtures/trivial_gem' }
